+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.2.2 - Unreleased
 * Remove `Console::print_cpu` and `Console::print_ram` methods.
 * Implement iNES mapper 1
+* Make `Console::step_frame` advance to the start of the next frame.
 
 ## 0.2.1 - June 6, 2019
 * First real release.